Add NH version specific profilers
