+++ source/Glimpse.NH/Glimpse.Nh.csproj
@@ -43,7 +43,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Plumbing\Injectors\WrapDbDriver.cs" />
-    <EmbeddedResource Include="Plumbing\Profiler\GlimpseProfileDbDriver.cs" />
+    <EmbeddedResource Include="Plumbing\Profiler\GlimpseProfileDbDriverNh320.cs" />
+    <EmbeddedResource Include="Plumbing\Profiler\GlimpseProfileDbDriverNh310.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Service\NH.cs" />
   </ItemGroup>
